requirements.txt
django>=1.9,<2.0

django-autoslug==1.9.3
jsonfield==1.0.3

djangorestframework==3.3.3

channels==0.11.0
asgi_redis==0.10.0